validation form login and register

+++ app/Http/Controllers/Api/TaskController.php
@@ -5,39 +5,42 @@ namespace App\Http\Controllers\Api;
 use App\Models\Task;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
+use Illuminate\Support\Facades\Auth;
 
 class TaskController extends BaseController
 {
     public function index()
     {
-       return Task::all();
+        $tasks = Auth::user()->tasks()->get();
+        return response()->json($tasks);
     }
 
     public function show($id)
     {
-        return Task::find($id);
+        return Auth::user()->tasks()->find($id);
     }
 
     public function store(StoreTaskRequest $request)
     {
-        $task = Task::create($request->all());
-
+        $task = Auth::user()->tasks()->create($request->all());
         return response()->json($task, 201);
     }
 
-    public function update(UpdateTaskRequest $request, $id)
+    public function update(UpdateTaskRequest $request, Task $task)
     {
-        $task = Task::findOrFail($id);
-        $task->update($request->all());
-
+        if (Auth::user()->id !== $task->user_id) {
+            return response()->json(['error' => 'You can only edit your own tasks.'], 403);
+        }
+        $task->update($request->only(['title']));
         return response()->json($task);
     }
 
-    public function destroy($id)
+    public function destroy(Task $task)
     {
-        $task = Task::findOrFail($id);
+        if (Auth::user()->id !== $task->user_id) {
+            return response()->json(['error' => 'You can only delete your own tasks.'], 403);
+        }
         $task->delete();
-
         return response()->json($task);
     }
 }
